2019.2.4

+++ C-third/第十九次作业/第十九次作业.cpp
@@ -5,7 +5,7 @@ char* Strncpy(char* destination, const char* source, size_t num)
 {
     assert(destination != NULL);
     assert(source != NULL);
-    int i = 0;
+    size_t i = 0;
     while(i < num && source[i] != '\0')
     {
         destination[i] = source[i];
@@ -29,12 +29,12 @@ char* Strncat(char* destination, const char* source, size_t num)
 {
     assert(destination != NULL);
     assert(source != NULL);
-    int i = 0;
+    size_t i = 0;
     while(destination[i] != '\0')
     {
         i++;
     }
-    int j = 0;
+    size_t j = 0;
     while(j < num && source[j] != '\0')
     {
         destination[i] = source[j];
@@ -49,7 +49,7 @@ int Strncmp(const char* str1, const char* str2, size_t num)
 {
     assert(str1 != NULL);
     assert(str2 != NULL);
-    int i = 0;
+    size_t i = 0;
     while(str1[i] == str2[i] && i < num && str1[i] != '\0' && str2[i] != '\0')
     {
         i++;
